docker/producer/requirements.txt
kafka-python
hdfs[client]
